Share single Spring context across sbs3 master tests
Introduce AbstractMasterContextTest, an abstract base that defines one
unioned @SpringBootTest context, and have the three non-endpoint master
tests extend it. Because the resolved context cache key is now identical
across them, the Spring TestContext framework caches and shares a single
ApplicationContext instead of building three.

AxelixMetadataEndpointTest is intentionally left untouched: its RANDOM_PORT
web server and JWT stack keep it in a separate context.

Verified with the spring-test-profiler report: the three tests share
context-8, while AxelixMetadataEndpointTest stays isolated in its own
context.
