+++ sbs/axelix-spring-boot-3-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/master/AbstractMasterContextTest.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright (C) 2025-2026 Axelix Labs
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.axelixlabs.axelix.sbs.spring.core.master;
+
+import java.lang.management.ManagementFactory;
+
+import org.springframework.boot.actuate.health.HealthEndpoint;
+import org.springframework.boot.autoconfigure.info.ProjectInfoAutoConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+
+import com.axelixlabs.axelix.common.api.registration.BasicDiscoveryMetadata;
+import com.axelixlabs.axelix.common.domain.version.AxelixVersionDiscoverer;
+
+/**
+ * Shared base for the {@code master} integration tests that do not require a running web server. By defining a single,
+ * unioned context here and having every subclass inherit it unchanged, the Spring TestContext framework resolves the
+ * same context cache key for all of them, so they share one cached {@link org.springframework.context.ApplicationContext}.
+ *
+ * <p>{@link AxelixMetadataEndpointTest} intentionally does not extend this class: it starts a real web server and its
+ * own JWT stack, so it keeps a separate context.
+ */
+@SpringBootTest
+@Import({
+    CommitIdPluginGitInformationProvider.class,
+    CommitIdPluginShortBuildInfoProvider.class,
+    ProjectInfoAutoConfiguration.class,
+    DefaultServiceMetadataAssembler.class,
+    AbstractMasterContextTest.SharedConfig.class
+})
+abstract class AbstractMasterContextTest {
+
+    @MockBean
+    private HealthEndpoint healthEndpoint;
+
+    @TestConfiguration
+    static class SharedConfig {
+
+        @Bean
+        VMFeaturesProvider vmFeaturesProvider() {
+            return new OptionsParsingVMFeaturesProvider(
+                    ManagementFactory.getRuntimeMXBean().getInputArguments());
+        }
+
+        @Bean
+        HealthDetectionFunction healthDetectionFunction() {
+            return () -> BasicDiscoveryMetadata.HealthStatus.UP;
+        }
+
+        @Bean
+        AxelixVersionDiscoverer axelixVersionDiscoverer() {
+            return () -> "1.1.3";
+        }
+
+        @Bean
+        public LibraryInformationProvider libraryInformationProvider() {
+            return new DefaultLibraryInformationProvider();
+        }
+    }
+}
